requirements.txt
boto3==1.9.117
botocore==1.12.117
dj-database-url==0.5.0
Django==2.1.7
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.9.1
django-storages==1.7.1
django-widget-tweaks==1.4.3
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
